Healthbars (bares de vida) y fix thow y Yerik contador

+++ Assets/Scripts/Invetory_ n_loot/item.cs
@@ -8,7 +8,7 @@ public class item : ScriptableObject
     public int id;
     public string itemName;
     public string description;
-    public int valueDMG;
+    public float valueDMG;
     public Sprite iconSprite;
     public string itemType;
     public GameObject Objeto;
